third_party/WebKit/LayoutTests/external/wpt/webrtc/interfaces-expected.txt

This is a testharness.js-based test.
Found 498 tests; 203 PASS, 295 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Main test driver
PASS Test driver for asyncInitCertificate
FAIL Test driver for asyncInitTransportsassert_unreached: Failed to run asyncInitTransports: OperationError: TEST_ERROR Reached unreachable code
FAIL Test driver for asyncInitMediaStreamTrackassert_unreached: Failed to run asyncInitMediaStreamTrack: NotSupportedError: Only secure origins are allowed (see: https://goo.gl/Y0ZkNV). Reached unreachable code
FAIL MediaStreamTrack interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS MediaStreamTrack interface object length
PASS MediaStreamTrack interface object name
FAIL MediaStreamTrack interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS MediaStreamTrack interface: existence and properties of interface prototype object's "constructor" property
FAIL MediaStreamTrack interface: attribute isolatedassert_true: The prototype object must have a property "isolated" expected true got false
FAIL MediaStreamTrack interface: attribute onisolationchangeassert_true: The prototype object must have a property "onisolationchange" expected true got false
FAIL MediaStreamTrack must be primary interface of idlTestObjects.mediaStreamTrackassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.mediaStreamTrackassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "isolated" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaStreamTrack interface: idlTestObjects.mediaStreamTrack must inherit property "onisolationchange" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL MediaStreamTrack must be primary interface of generateMediaStreamTrack('audio')assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL Stringification of generateMediaStreamTrack('audio')assert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "isolated" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL MediaStreamTrack interface: generateMediaStreamTrack('audio') must inherit property "onisolationchange" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "Error: assert_idl_attribute: Expect pc to have addTransceiver() method property "addTransceiver" not found in prototype chain"
FAIL RTCPeerConnection interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnection interface object length
PASS RTCPeerConnection interface object name
FAIL RTCPeerConnection interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnection interface: existence and properties of interface prototype object's "constructor" property
PASS RTCPeerConnection interface: operation createOffer(RTCOfferOptions)
PASS RTCPeerConnection interface: operation createAnswer(RTCAnswerOptions)
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit)
PASS RTCPeerConnection interface: attribute localDescription
FAIL RTCPeerConnection interface: attribute currentLocalDescriptionassert_true: The prototype object must have a property "currentLocalDescription" expected true got false
FAIL RTCPeerConnection interface: attribute pendingLocalDescriptionassert_true: The prototype object must have a property "pendingLocalDescription" expected true got false
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit)
PASS RTCPeerConnection interface: attribute remoteDescription
FAIL RTCPeerConnection interface: attribute currentRemoteDescriptionassert_true: The prototype object must have a property "currentRemoteDescription" expected true got false
FAIL RTCPeerConnection interface: attribute pendingRemoteDescriptionassert_true: The prototype object must have a property "pendingRemoteDescription" expected true got false
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object])
PASS RTCPeerConnection interface: attribute signalingState
PASS RTCPeerConnection interface: attribute iceGatheringState
PASS RTCPeerConnection interface: attribute iceConnectionState
FAIL RTCPeerConnection interface: attribute connectionStateassert_true: The prototype object must have a property "connectionState" expected true got false
FAIL RTCPeerConnection interface: attribute canTrickleIceCandidatesassert_true: The prototype object must have a property "canTrickleIceCandidates" expected true got false
FAIL RTCPeerConnection interface: operation getDefaultIceServers()assert_own_property: interface object missing static operation expected property "getDefaultIceServers" missing
FAIL RTCPeerConnection interface: operation getConfiguration()assert_own_property: interface prototype object missing non-static operation expected property "getConfiguration" missing
PASS RTCPeerConnection interface: operation setConfiguration(RTCConfiguration)
PASS RTCPeerConnection interface: operation close()
PASS RTCPeerConnection interface: attribute onnegotiationneeded
PASS RTCPeerConnection interface: attribute onicecandidate
FAIL RTCPeerConnection interface: attribute onicecandidateerrorassert_true: The prototype object must have a property "onicecandidateerror" expected true got false
PASS RTCPeerConnection interface: attribute onsignalingstatechange
PASS RTCPeerConnection interface: attribute oniceconnectionstatechange
PASS RTCPeerConnection interface: attribute onicegatheringstatechange
FAIL RTCPeerConnection interface: attribute onconnectionstatechangeassert_true: The prototype object must have a property "onconnectionstatechange" expected true got false
FAIL RTCPeerConnection interface: attribute onfingerprintfailureassert_true: The prototype object must have a property "onfingerprintfailure" expected true got false
PASS RTCPeerConnection interface: operation createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions)
PASS RTCPeerConnection interface: operation setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)
PASS RTCPeerConnection interface: operation createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback)
PASS RTCPeerConnection interface: operation setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)
PASS RTCPeerConnection interface: operation addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback)
PASS RTCPeerConnection interface: operation generateCertificate(AlgorithmIdentifier)
PASS RTCPeerConnection interface: operation getSenders()
PASS RTCPeerConnection interface: operation getReceivers()
FAIL RTCPeerConnection interface: operation getTransceivers()assert_own_property: interface prototype object missing non-static operation expected property "getTransceivers" missing
PASS RTCPeerConnection interface: operation addTrack(MediaStreamTrack, MediaStream)
PASS RTCPeerConnection interface: operation removeTrack(RTCRtpSender)
FAIL RTCPeerConnection interface: operation addTransceiver([object Object],[object Object], RTCRtpTransceiverInit)assert_own_property: interface prototype object missing non-static operation expected property "addTransceiver" missing
PASS RTCPeerConnection interface: attribute ontrack
FAIL RTCPeerConnection interface: attribute sctpassert_true: The prototype object must have a property "sctp" expected true got false
PASS RTCPeerConnection interface: operation createDataChannel(USVString, RTCDataChannelInit)
PASS RTCPeerConnection interface: attribute ondatachannel
PASS RTCPeerConnection interface: operation getStats(MediaStreamTrack)
FAIL RTCPeerConnection interface: operation setIdentityProvider(DOMString, RTCIdentityProviderOptions)assert_own_property: interface prototype object missing non-static operation expected property "setIdentityProvider" missing
FAIL RTCPeerConnection interface: operation getIdentityAssertion()assert_own_property: interface prototype object missing non-static operation expected property "getIdentityAssertion" missing
FAIL RTCPeerConnection interface: attribute peerIdentityassert_true: The prototype object must have a property "peerIdentity" expected true got false
FAIL RTCPeerConnection interface: attribute idpLoginUrlassert_true: The prototype object must have a property "idpLoginUrl" expected true got false
FAIL RTCPeerConnection interface: attribute idpErrorInfoassert_true: The prototype object must have a property "idpErrorInfo" expected true got false
PASS RTCPeerConnection must be primary interface of new RTCPeerConnection()
FAIL Stringification of new RTCPeerConnection()Cannot read property 'has_stringifier' of undefined
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCOfferOptions)" with the proper type
PASS RTCPeerConnection interface: calling createOffer(RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCAnswerOptions)" with the proper type
PASS RTCPeerConnection interface: calling createAnswer(RTCAnswerOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit)" with the proper type
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "localDescription" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentLocalDescription" with the proper typeassert_inherits: property "currentLocalDescription" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingLocalDescription" with the proper typeassert_inherits: property "pendingLocalDescription" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit)" with the proper type
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "remoteDescription" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "currentRemoteDescription" with the proper typeassert_inherits: property "currentRemoteDescription" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "pendingRemoteDescription" with the proper typeassert_inherits: property "pendingRemoteDescription" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object])" with the proper type
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object]) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "signalingState" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceGatheringState" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "iceConnectionState" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "connectionState" with the proper typeassert_inherits: property "connectionState" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "canTrickleIceCandidates" with the proper typeassert_inherits: property "canTrickleIceCandidates" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getDefaultIceServers()" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getConfiguration()" with the proper typeassert_inherits: property "getConfiguration" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setConfiguration(RTCConfiguration)" with the proper type
PASS RTCPeerConnection interface: calling setConfiguration(RTCConfiguration) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "close()" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onnegotiationneeded" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidate" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicecandidateerror" with the proper typeassert_inherits: property "onicecandidateerror" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onsignalingstatechange" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "oniceconnectionstatechange" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onicegatheringstatechange" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onconnectionstatechange" with the proper typeassert_inherits: property "onconnectionstatechange" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "onfingerprintfailure" with the proper typeassert_inherits: property "onfingerprintfailure" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions)" with the proper type
PASS RTCPeerConnection interface: calling createOffer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback, RTCOfferOptions) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling setLocalDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling createAnswer(RTCSessionDescriptionCallback, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling setRemoteDescription(RTCSessionDescriptionInit, VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback)" with the proper type
PASS RTCPeerConnection interface: calling addIceCandidate([object Object],[object Object], VoidFunction, RTCPeerConnectionErrorCallback) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "generateCertificate(AlgorithmIdentifier)" with the proper type
PASS RTCPeerConnection interface: calling generateCertificate(AlgorithmIdentifier) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getSenders()" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getReceivers()" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getTransceivers()" with the proper typeassert_inherits: property "getTransceivers" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTrack(MediaStreamTrack, MediaStream)" with the proper type
PASS RTCPeerConnection interface: calling addTrack(MediaStreamTrack, MediaStream) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "removeTrack(RTCRtpSender)" with the proper type
PASS RTCPeerConnection interface: calling removeTrack(RTCRtpSender) on new RTCPeerConnection() with too few arguments must throw TypeError
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "addTransceiver([object Object],[object Object], RTCRtpTransceiverInit)" with the proper typeassert_inherits: property "addTransceiver" not found in prototype chain
FAIL RTCPeerConnection interface: calling addTransceiver([object Object],[object Object], RTCRtpTransceiverInit) on new RTCPeerConnection() with too few arguments must throw TypeErrorassert_inherits: property "addTransceiver" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ontrack" with the proper type
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "sctp" with the proper typeassert_inherits: property "sctp" not found in prototype chain
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "createDataChannel(USVString, RTCDataChannelInit)" with the proper type
PASS RTCPeerConnection interface: calling createDataChannel(USVString, RTCDataChannelInit) on new RTCPeerConnection() with too few arguments must throw TypeError
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "ondatachannel" with the proper type
PASS RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getStats(MediaStreamTrack)" with the proper type
PASS RTCPeerConnection interface: calling getStats(MediaStreamTrack) on new RTCPeerConnection() with too few arguments must throw TypeError
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "setIdentityProvider(DOMString, RTCIdentityProviderOptions)" with the proper typeassert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: calling setIdentityProvider(DOMString, RTCIdentityProviderOptions) on new RTCPeerConnection() with too few arguments must throw TypeErrorassert_inherits: property "setIdentityProvider" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "getIdentityAssertion()" with the proper typeassert_inherits: property "getIdentityAssertion" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "peerIdentity" with the proper typeassert_inherits: property "peerIdentity" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpLoginUrl" with the proper typeassert_inherits: property "idpLoginUrl" not found in prototype chain
FAIL RTCPeerConnection interface: new RTCPeerConnection() must inherit property "idpErrorInfo" with the proper typeassert_inherits: property "idpErrorInfo" not found in prototype chain
PASS RTCSessionDescription interface: existence and properties of interface object
FAIL RTCSessionDescription interface object lengthassert_equals: wrong value for RTCSessionDescription.length expected 1 but got 0
PASS RTCSessionDescription interface object name
PASS RTCSessionDescription interface: existence and properties of interface prototype object
PASS RTCSessionDescription interface: existence and properties of interface prototype object's "constructor" property
FAIL RTCSessionDescription interface: attribute typeassert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
FAIL RTCSessionDescription interface: attribute sdpassert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
PASS RTCSessionDescription must be primary interface of new RTCSessionDescription({ type: 'offer' })
PASS Stringification of new RTCSessionDescription({ type: 'offer' })
PASS RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "type" with the proper type
FAIL RTCSessionDescription interface: new RTCSessionDescription({ type: 'offer' }) must inherit property "sdp" with the proper typeassert_equals: expected "string" but got "object"
PASS RTCIceCandidate interface: existence and properties of interface object
FAIL RTCIceCandidate interface object lengthassert_equals: wrong value for RTCIceCandidate.length expected 0 but got 1
PASS RTCIceCandidate interface object name
PASS RTCIceCandidate interface: existence and properties of interface prototype object
PASS RTCIceCandidate interface: existence and properties of interface prototype object's "constructor" property
FAIL RTCIceCandidate interface: attribute candidateassert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
FAIL RTCIceCandidate interface: attribute sdpMidassert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
FAIL RTCIceCandidate interface: attribute sdpMLineIndexassert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function () { [native code] }"
FAIL RTCIceCandidate interface: attribute foundationassert_true: The prototype object must have a property "foundation" expected true got false
FAIL RTCIceCandidate interface: attribute componentassert_true: The prototype object must have a property "component" expected true got false
FAIL RTCIceCandidate interface: attribute priorityassert_true: The prototype object must have a property "priority" expected true got false
FAIL RTCIceCandidate interface: attribute ipassert_true: The prototype object must have a property "ip" expected true got false
FAIL RTCIceCandidate interface: attribute protocolassert_true: The prototype object must have a property "protocol" expected true got false
FAIL RTCIceCandidate interface: attribute portassert_true: The prototype object must have a property "port" expected true got false
FAIL RTCIceCandidate interface: attribute typeassert_true: The prototype object must have a property "type" expected true got false
FAIL RTCIceCandidate interface: attribute tcpTypeassert_true: The prototype object must have a property "tcpType" expected true got false
FAIL RTCIceCandidate interface: attribute relatedAddressassert_true: The prototype object must have a property "relatedAddress" expected true got false
FAIL RTCIceCandidate interface: attribute relatedPortassert_true: The prototype object must have a property "relatedPort" expected true got false
FAIL RTCIceCandidate interface: attribute ufragassert_true: The prototype object must have a property "ufrag" expected true got false
FAIL RTCIceCandidate must be primary interface of new RTCIceCandidate({ sdpMid: 1 })assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL Stringification of new RTCIceCandidate({ sdpMid: 1 })assert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "candidate" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMid" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "sdpMLineIndex" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "foundation" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "component" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "priority" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "ip" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "protocol" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "port" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "type" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "tcpType" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedAddress" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "relatedPort" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCIceCandidate interface: new RTCIceCandidate({ sdpMid: 1 }) must inherit property "ufrag" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty."
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnectionIceEvent interface object length
PASS RTCPeerConnectionIceEvent interface object name
FAIL RTCPeerConnectionIceEvent interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS RTCPeerConnectionIceEvent interface: existence and properties of interface prototype object's "constructor" property
PASS RTCPeerConnectionIceEvent interface: attribute candidate
FAIL RTCPeerConnectionIceEvent interface: attribute urlassert_true: The prototype object must have a property "url" expected true got false
PASS RTCPeerConnectionIceEvent must be primary interface of new RTCPeerConnectionIceEvent('ice')
FAIL Stringification of new RTCPeerConnectionIceEvent('ice')Cannot read property 'has_stringifier' of undefined
PASS RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "candidate" with the proper type
FAIL RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "url" with the proper typeassert_inherits: property "url" not found in prototype chain
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object lengthassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface object nameassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute hostCandidateassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute urlassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorCodeassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent interface: attribute errorTextassert_own_property: self does not have own property "RTCPeerConnectionIceErrorEvent" expected property "RTCPeerConnectionIceErrorEvent" missing
FAIL RTCPeerConnectionIceErrorEvent must be primary interface of new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 });assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
FAIL Stringification of new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 });assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "hostCandidate" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "url" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "errorCode" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
FAIL RTCPeerConnectionIceErrorEvent interface: new RTCPeerConnectionIceErrorEvent('ice-error', { errorCode: 701 }); must inherit property "errorText" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCPeerConnectionIceErrorEvent is not defined"
PASS RTCCertificate interface: existence and properties of interface object
PASS RTCCertificate interface object length
PASS RTCCertificate interface object name
PASS RTCCertificate interface: existence and properties of interface prototype object
PASS RTCCertificate interface: existence and properties of interface prototype object's "constructor" property
PASS RTCCertificate interface: attribute expires
PASS RTCCertificate interface: operation getFingerprints()
FAIL RTCCertificate interface: operation getAlgorithm()assert_own_property: interface prototype object missing non-static operation expected property "getAlgorithm" missing
PASS RTCCertificate must be primary interface of idlTestObjects.certificate
PASS Stringification of idlTestObjects.certificate
PASS RTCCertificate interface: idlTestObjects.certificate must inherit property "expires" with the proper type
PASS RTCCertificate interface: idlTestObjects.certificate must inherit property "getFingerprints()" with the proper type
FAIL RTCCertificate interface: idlTestObjects.certificate must inherit property "getAlgorithm()" with the proper typeassert_inherits: property "getAlgorithm" not found in prototype chain
PASS RTCRtpSender interface: existence and properties of interface object
PASS RTCRtpSender interface object length
PASS RTCRtpSender interface object name
PASS RTCRtpSender interface: existence and properties of interface prototype object
PASS RTCRtpSender interface: existence and properties of interface prototype object's "constructor" property
PASS RTCRtpSender interface: attribute track
FAIL RTCRtpSender interface: attribute transportassert_true: The prototype object must have a property "transport" expected true got false
FAIL RTCRtpSender interface: attribute rtcpTransportassert_true: The prototype object must have a property "rtcpTransport" expected true got false
FAIL RTCRtpSender interface: operation getCapabilities(DOMString)assert_own_property: interface object missing static operation expected property "getCapabilities" missing
FAIL RTCRtpSender interface: operation setParameters(RTCRtpParameters)assert_own_property: interface prototype object missing non-static operation expected property "setParameters" missing
FAIL RTCRtpSender interface: operation getParameters()assert_own_property: interface prototype object missing non-static operation expected property "getParameters" missing
FAIL RTCRtpSender interface: operation replaceTrack(MediaStreamTrack)assert_own_property: interface prototype object missing non-static operation expected property "replaceTrack" missing
FAIL RTCRtpSender interface: operation getStats()assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
FAIL RTCRtpSender interface: attribute dtmfassert_true: The prototype object must have a property "dtmf" expected true got false
FAIL RTCRtpSender must be primary interface of new RTCPeerConnection().addTransceiver('audio').senderassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL Stringification of new RTCPeerConnection().addTransceiver('audio').senderassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "track" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "transport" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "rtcpTransport" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getCapabilities(DOMString)" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: calling getCapabilities(DOMString) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "setParameters(RTCRtpParameters)" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: calling setParameters(RTCRtpParameters) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getParameters()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "replaceTrack(MediaStreamTrack)" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: calling replaceTrack(MediaStreamTrack) on new RTCPeerConnection().addTransceiver('audio').sender with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "getStats()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpSender interface: new RTCPeerConnection().addTransceiver('audio').sender must inherit property "dtmf" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
PASS RTCRtpReceiver interface: existence and properties of interface object
PASS RTCRtpReceiver interface object length
PASS RTCRtpReceiver interface object name
PASS RTCRtpReceiver interface: existence and properties of interface prototype object
PASS RTCRtpReceiver interface: existence and properties of interface prototype object's "constructor" property
PASS RTCRtpReceiver interface: attribute track
FAIL RTCRtpReceiver interface: attribute transportassert_true: The prototype object must have a property "transport" expected true got false
FAIL RTCRtpReceiver interface: attribute rtcpTransportassert_true: The prototype object must have a property "rtcpTransport" expected true got false
FAIL RTCRtpReceiver interface: operation getCapabilities(DOMString)assert_own_property: interface object missing static operation expected property "getCapabilities" missing
FAIL RTCRtpReceiver interface: operation getParameters()assert_own_property: interface prototype object missing non-static operation expected property "getParameters" missing
PASS RTCRtpReceiver interface: operation getContributingSources()
FAIL RTCRtpReceiver interface: operation getSynchronizationSources()assert_own_property: interface prototype object missing non-static operation expected property "getSynchronizationSources" missing
FAIL RTCRtpReceiver interface: operation getStats()assert_own_property: interface prototype object missing non-static operation expected property "getStats" missing
FAIL RTCRtpReceiver must be primary interface of new RTCPeerConnection().addTransceiver('audio').receiverassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL Stringification of new RTCPeerConnection().addTransceiver('audio').receiverassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "track" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "transport" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "rtcpTransport" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getCapabilities(DOMString)" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: calling getCapabilities(DOMString) on new RTCPeerConnection().addTransceiver('audio').receiver with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getParameters()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getContributingSources()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getSynchronizationSources()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "getStats()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
PASS RTCRtpContributingSource interface: existence and properties of interface object
PASS RTCRtpContributingSource interface object length
PASS RTCRtpContributingSource interface object name
PASS RTCRtpContributingSource interface: existence and properties of interface prototype object
PASS RTCRtpContributingSource interface: existence and properties of interface prototype object's "constructor" property
PASS RTCRtpContributingSource interface: attribute timestamp
PASS RTCRtpContributingSource interface: attribute source
FAIL RTCRtpContributingSource interface: attribute audioLevelassert_true: The prototype object must have a property "audioLevel" expected true got false
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object lengthassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface object nameassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute timestampassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute sourceassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute audioLevelassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpSynchronizationSource interface: attribute voiceActivityFlagassert_own_property: self does not have own property "RTCRtpSynchronizationSource" expected property "RTCRtpSynchronizationSource" missing
FAIL RTCRtpTransceiver interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface object lengthassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface object nameassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute midassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute senderassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute receiverassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute stoppedassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute directionassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: attribute currentDirectionassert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: operation setDirection(RTCRtpTransceiverDirection)assert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: operation stop()assert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver interface: operation setCodecPreferences([object Object])assert_own_property: self does not have own property "RTCRtpTransceiver" expected property "RTCRtpTransceiver" missing
FAIL RTCRtpTransceiver must be primary interface of new RTCPeerConnection().addTransceiver('audio')assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL Stringification of new RTCPeerConnection().addTransceiver('audio')assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "mid" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "sender" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "receiver" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "stopped" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "direction" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "currentDirection" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "setDirection(RTCRtpTransceiverDirection)" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: calling setDirection(RTCRtpTransceiverDirection) on new RTCPeerConnection().addTransceiver('audio') with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "stop()" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: new RTCPeerConnection().addTransceiver('audio') must inherit property "setCodecPreferences([object Object])" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCRtpTransceiver interface: calling setCodecPreferences([object Object]) on new RTCPeerConnection().addTransceiver('audio') with too few arguments must throw TypeErrorassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (intermediate value).addTransceiver is not a function"
FAIL RTCDtlsTransport interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object lengthassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface object nameassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute transportassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute stateassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: operation getRemoteCertificates()assert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport interface: attribute onstatechangeassert_own_property: self does not have own property "RTCDtlsTransport" expected property "RTCDtlsTransport" missing
FAIL RTCDtlsTransport must be primary interface of idlTestObjects.dtlsTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.dtlsTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "transport" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "state" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "getRemoteCertificates()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDtlsTransport interface: idlTestObjects.dtlsTransport must inherit property "onstatechange" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface object lengthassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface object nameassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute roleassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute componentassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute stateassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute gatheringStateassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: operation getLocalCandidates()assert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: operation getRemoteCandidates()assert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: operation getSelectedCandidatePair()assert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: operation getLocalParameters()assert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: operation getRemoteParameters()assert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute onstatechangeassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute ongatheringstatechangeassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport interface: attribute onselectedcandidatepairchangeassert_own_property: self does not have own property "RTCIceTransport" expected property "RTCIceTransport" missing
FAIL RTCIceTransport must be primary interface of idlTestObjects.iceTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.iceTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "role" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "component" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "state" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "gatheringState" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getLocalCandidates()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getRemoteCandidates()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getSelectedCandidatePair()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getLocalParameters()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "getRemoteParameters()" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "onstatechange" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "ongatheringstatechange" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "onselectedcandidatepairchange" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCTrackEvent interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS RTCTrackEvent interface object length
PASS RTCTrackEvent interface object name
FAIL RTCTrackEvent interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS RTCTrackEvent interface: existence and properties of interface prototype object's "constructor" property
PASS RTCTrackEvent interface: attribute receiver
PASS RTCTrackEvent interface: attribute track
PASS RTCTrackEvent interface: attribute streams
FAIL RTCTrackEvent interface: attribute transceiverassert_true: The prototype object must have a property "transceiver" expected true got false
FAIL RTCTrackEvent must be primary interface of initTrackEvent()assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL Stringification of initTrackEvent()assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL RTCTrackEvent interface: initTrackEvent() must inherit property "receiver" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL RTCTrackEvent interface: initTrackEvent() must inherit property "track" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL RTCTrackEvent interface: initTrackEvent() must inherit property "streams" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL RTCTrackEvent interface: initTrackEvent() must inherit property "transceiver" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: pc.addTransceiver is not a function"
FAIL RTCSctpTransport interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object lengthassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface object nameassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: attribute transportassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport interface: attribute maxMessageSizeassert_own_property: self does not have own property "RTCSctpTransport" expected property "RTCSctpTransport" missing
FAIL RTCSctpTransport must be primary interface of idlTestObjects.sctpTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of idlTestObjects.sctpTransportassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCSctpTransport interface: idlTestObjects.sctpTransport must inherit property "transport" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCSctpTransport interface: idlTestObjects.sctpTransport must inherit property "maxMessageSize" with the proper typeassert_equals: wrong typeof object expected "object" but got "undefined"
FAIL RTCDataChannel interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS RTCDataChannel interface object length
PASS RTCDataChannel interface object name
FAIL RTCDataChannel interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS RTCDataChannel interface: existence and properties of interface prototype object's "constructor" property
PASS RTCDataChannel interface: attribute label
PASS RTCDataChannel interface: attribute ordered
FAIL RTCDataChannel interface: attribute maxPacketLifeTimeassert_true: The prototype object must have a property "maxPacketLifeTime" expected true got false
PASS RTCDataChannel interface: attribute maxRetransmits
PASS RTCDataChannel interface: attribute protocol
PASS RTCDataChannel interface: attribute negotiated
PASS RTCDataChannel interface: attribute id
FAIL RTCDataChannel interface: attribute priorityassert_true: The prototype object must have a property "priority" expected true got false
PASS RTCDataChannel interface: attribute readyState
PASS RTCDataChannel interface: attribute bufferedAmount
PASS RTCDataChannel interface: attribute bufferedAmountLowThreshold
PASS RTCDataChannel interface: attribute onopen
PASS RTCDataChannel interface: attribute onbufferedamountlow
PASS RTCDataChannel interface: attribute onerror
PASS RTCDataChannel interface: attribute onclose
PASS RTCDataChannel interface: operation close()
PASS RTCDataChannel interface: attribute onmessage
PASS RTCDataChannel interface: attribute binaryType
PASS RTCDataChannel interface: operation send(USVString)
PASS RTCDataChannel interface: operation send(Blob)
PASS RTCDataChannel interface: operation send(ArrayBuffer)
PASS RTCDataChannel interface: operation send(ArrayBufferView)
PASS RTCDataChannel must be primary interface of new RTCPeerConnection().createDataChannel('')
FAIL Stringification of new RTCPeerConnection().createDataChannel('')Cannot read property 'has_stringifier' of undefined
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "label" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "ordered" with the proper type
FAIL RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "maxPacketLifeTime" with the proper typeassert_inherits: property "maxPacketLifeTime" not found in prototype chain
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "maxRetransmits" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "protocol" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "negotiated" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "id" with the proper type
FAIL RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "priority" with the proper typeassert_inherits: property "priority" not found in prototype chain
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "readyState" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "bufferedAmount" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "bufferedAmountLowThreshold" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onopen" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onbufferedamountlow" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onerror" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onclose" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "close()" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "onmessage" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "binaryType" with the proper type
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(USVString)" with the proper type
PASS RTCDataChannel interface: calling send(USVString) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(Blob)" with the proper type
PASS RTCDataChannel interface: calling send(Blob) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(ArrayBuffer)" with the proper type
PASS RTCDataChannel interface: calling send(ArrayBuffer) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError
PASS RTCDataChannel interface: new RTCPeerConnection().createDataChannel('') must inherit property "send(ArrayBufferView)" with the proper type
PASS RTCDataChannel interface: calling send(ArrayBufferView) on new RTCPeerConnection().createDataChannel('') with too few arguments must throw TypeError
FAIL RTCDataChannelEvent interface: existence and properties of interface objectCannot read property 'has_extended_attribute' of undefined
PASS RTCDataChannelEvent interface object length
PASS RTCDataChannelEvent interface object name
FAIL RTCDataChannelEvent interface: existence and properties of interface prototype objectCannot read property 'has_extended_attribute' of undefined
PASS RTCDataChannelEvent interface: existence and properties of interface prototype object's "constructor" property
PASS RTCDataChannelEvent interface: attribute channel
PASS RTCDataChannelEvent must be primary interface of new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') })
FAIL Stringification of new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') })Cannot read property 'has_stringifier' of undefined
PASS RTCDataChannelEvent interface: new RTCDataChannelEvent('channel',
        { channel: new RTCPeerConnection().createDataChannel('') }) must inherit property "channel" with the proper type
FAIL RTCDTMFSender interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object lengthassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface object nameassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: operation insertDTMF(DOMString, unsigned long, unsigned long)assert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: attribute ontonechangeassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFSender interface: attribute toneBufferassert_own_property: self does not have own property "RTCDTMFSender" expected property "RTCDTMFSender" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object lengthassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface object nameassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
FAIL RTCDTMFToneChangeEvent interface: attribute toneassert_own_property: self does not have own property "RTCDTMFToneChangeEvent" expected property "RTCDTMFToneChangeEvent" missing
PASS RTCStatsReport interface: existence and properties of interface object
PASS RTCStatsReport interface object length
PASS RTCStatsReport interface object name
PASS RTCStatsReport interface: existence and properties of interface prototype object
PASS RTCStatsReport interface: existence and properties of interface prototype object's "constructor" property
PASS RTCIdentityProviderGlobalScope interface: existence and properties of interface object
PASS RTCIdentityProviderRegistrar interface: existence and properties of interface object
FAIL RTCIdentityAssertion interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object lengthassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface object nameassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: attribute idpassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCIdentityAssertion interface: attribute nameassert_own_property: self does not have own property "RTCIdentityAssertion" expected property "RTCIdentityAssertion" missing
FAIL RTCErrorEvent interface: existence and properties of interface objectassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object lengthassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface object nameassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype objectassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: existence and properties of interface prototype object's "constructor" propertyassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent interface: attribute errorassert_own_property: self does not have own property "RTCErrorEvent" expected property "RTCErrorEvent" missing
FAIL RTCErrorEvent must be primary interface of new RTCErrorEvent('error')assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCErrorEvent is not defined"
FAIL Stringification of new RTCErrorEvent('error')assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCErrorEvent is not defined"
FAIL RTCErrorEvent interface: new RTCErrorEvent('error') must inherit property "error" with the proper typeassert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: RTCErrorEvent is not defined"
Harness: the test ran to completion.
